Progress was made with the final sequencer app.

+++ Sequencer/include/FmOscillator.h
@@ -6,7 +6,7 @@ class FmOscillator : public Oscillator {
 public:
 	FmOscillator( int sampleRate, float frequency,
 			float harmonicity, float modulationIndex );
-	virtual short nextSample();
+    virtual float nextSample();
 private:
 	float m_harmonicity;
 	float m_modulationIndex;
