+++ Sequencer/include/RisingSawtoothOscillator.h
@@ -6,7 +6,7 @@ class RisingSawtoothOscillator : public Oscillator {
 public:
 	explicit RisingSawtoothOscillator( float frequency = 440.0f );
 	RisingSawtoothOscillator( int sampleRate, float frequency );
-	virtual short nextSample();
+    virtual float nextSample();
 };
 
 #endif // RISINGSAWTOOTHOSCILLATOR_H
